Set lmxo always to the maximum l of the basis orbitals
When perturbative polarization orbitals are present, the
maximum l reported for the basis set (lmxo) could sometimes
be underestimated (only the 'working' shells were taken into
account). Now lmxo is set always to the real maximum.

In case a highest-l shell is 'empty', it is reported as
a polarization orbital.