Enabled LDOS and PDOS energy specification with respect to EF
Converted projected_DOS to F90.

Documented the new DOS.kgrid.* and L/PDOS.kgrid.* input options.

Now LDOS and PDOS may be prefixed with EF to specify that the energies
are with respect to the Fermi level. This may help users who want to use the same input for various systems.