Fixed NCDF write-out for MD and FC calculations
Previously the NCDF file-write-out would crash when performing
MD/FC calculations because the sparse arrays does not really make sense
to store.

Now the file content is based on the type of run, single/MD/FC.
A relaxation is equivalent to an MD (currently).

One non-CDF related change is the change of MD.NumCGSteps to
MD.Steps keyword. The NumCGSteps is deprecated since many more
options have entered and CG was not necessarily the correct instance.

Secondly, the MD.Steps keyword is now also the default for
MD.FinalTimeStep which means that a single keyword determines the number
of steps runned.

The old flags are still used as the default values.