Bumped nmesh to 1024 in m_filter.f90
Certain FilterCutoff values did not result in usable
basis functions. This change should resolve most of the problems.

This result will change all basis-functions using the FilterCutoff option
and destroys any backward compatibility for such calculations. Since
the filtering is only applied when FilterCutoff is specified the majority
of calculations should be unaffected.