Bugfix in the copy_supercell routine
Also added a new subroutine to correct a supercell_sp_d2 in-place.
This 2nd routine will change the column values to the new supercell places
and if the new supercell does not exist the value will be deleted.
A further reduction of the sparse pattern must happen in another
routine.