Removal of some non-standard constructs
In order to remain compatible with the Fortran2003 standard (the current
target in Siesta), some constructs have been removed.

Compiling with gfortran with '-std=f2003’ found a few issues:

— use of ‘newunit=‘ (f2008 feature)

— naked use of 'derf' in fermid (cold-smearing patch) — erf is std only
  in f2008, and derf is a GNU extension.

These have been corrected in this patch (the latter by 'use'ing the
already existing m_errorf module).

Then, the compiler flagged some more errors, having to do with the use
of GNU extensions:

  dcmplx, dreal, dconjg, dimag, and cdabs

Only a limited number of dcmplx appearances have been fixed in this
patch. The rest will be dealt with in the near future.