Added doxygen creation (using Ford files)
The documentation created by doxygen is now very limited.
However, it may be used to check that code-documentation is
ford-doxygen compatible!
Indeed the sparse_matrices.f90

There are some major difficulties in implementation for both:

- ford accepts markdown-like links with specific entries for the
  pages site, i.e.: [see here](|page|/.../page.html).
  These are not used by Doxygen
- Doxygen uses \* commands for creating specific entries etc.
  These are not used by ford

Bottomline, continue using FORD, but iff documentation is added
in the code files, check that it runs using Doxygen.