Change the spin-reduction convention for entropy
For ELSI versions after 2018-08-17, there is no need to reduce over spins
when computing the entropy.

To use older versions (such as the 2.0.2 release), use the preprocessor symbol

    SIESTA__ELSI__OLD_SPIN_CONVENTION

If the version is 'old', but the above symbol is not used, a compilation error will
be triggered. (This works by means of a call to 'elsi_get_datestamp', which was added just
around the same time as the spin-reduction-convention change: Aug 17 vs Aug 21, 2018.
ELSI versions in between will not work without further intervention.