Bugfix when using elec-box in the Hartree solution and Bloch-expansion
There were problems when using the Bloch expansion and a box-Hartree/Voltage
solution. Effectively the electrode box was only the minimal box which meant
that it failed to capture the repeated atoms and thus would crash.

Now the expansion of the box is fixed and the suggestions for displacing the
device is better.

Thanks to Gaetano Calogero for finding the bug.

Also added a flag to explicitly specify which electrode should be used as the
electrode-plane.