Implemented a Gamma-only sparse pattern extrapolator
When starting from a DM originating from a Gamma-only calculation the only elements that
exists are elements that "looks" like they exist in the primary unit cell. However,
they in fact correspond to supercell connections. Now the code will try and unfold
the Gamma-only calculation DM elements to the supercell format by copying the elements
multiple times (this comes from the fact that the DM is a Gamma-only with auxiliary cell
has the same DM elements for all orbital connections).

There are cases where this does not hold, e.g. when users create their own initial DM, in those
cases one can use the flag: DM.Init.Unfold false to disable the unfolding.