Implemented auxiliary cell to Gamma-only sparse pattern folding mechanism
When starting from a DM originating from a supercell calculation, and
performing a Gamma-only calculation, then connections may be doubled when
io,jo == io,io where jo is the supercell orbital.
This should equate to summing the DM contributions for all equal orbitals.

This is now implemented and tested for the si64 system
1. no auxiliary cell
2. auxiliary, init with DM from 1.
3. no auxiliary, init with DM from 2.