Applied patch for fixing folding into 000 supercells
This change has surfaced from discussion with Alberto (his fix).
This change stores the 000 element in the supercell DM, and if
this element does not exist it will store a "random" supercell DM
element. This has some problems for very small unit-cells, however,
we expect this problem to be insignificant compared to the problems it
solves.