Reduced disk-space usage for Gamma-only electrodes
A huge memory reduction can be done when an electrode is transverse k-point Gamma-only.
I.e. if there are no interactions along the transverse directions there
is no point in re-calculating the same self-energy over and over.

This is often encountered in N>2 terminal calculations and in e.g. graphene-GNR
constrictions.

This feature is not fully tested and should be tested before 4.1-b4 is released.