Allowed +/- for real-space SE
When using the real-space SE one still needs planes to fix the
Hartree potential. Prior to this commit the sign was fixed to '-'
which meant that only one side of the cell-box was checked for the
Hartree plane correction. However, the user should decide because
one atomic plane may lie on the cell-edge.

Also, a few places we calculated the area of two spanning vectors.
The equations are now fixed by using the cross-product. The prior
used equations only worked correctly for orthogonal cells.