Removed Metis pivoting in TranSiesta analyze
It *always* produced bad pivoting, and for disconnected graphs
it would crash.
Now one can enable it using a pre-processor statement:
  TS_PVT_METIS
which is not documented because I don't think it is even applicable.
So better not confuse users to think it is better by having it optionally.

The timer printed for the analyzation is now also for the analyzation method.