Fixed grid-cell sampling and output density, lp:1799991
The output density is not according to the input atomic coordinates.
This was because the neighbouring lists was only updated for the last
displaced atoms.
The fix is however very easy since we simply call the dhscf_init
routine again with the input atomic coordinates after having performed
the grid-cell sampling.