Fixed some more tests with few core requirements, also discovered a bug in OMM
This partially fixes the OMM bug since the psi array wasn't allocated on entry.
The OMM Gamma-only calculation is still faulty because the psi array *should*
contain the previous diagon wavefunction coefficients. This commit simply
ensures the arrays to be allocated. But with wrong numbers. I.e. the subsequent
OMM cycle produces NaN because psi is 0.