Fix OMM reuse of the psi array from 'densematrix' module
The OMM method can use eigenvectors from diagonalization as seeds. In
that case the first few nominal OMM iterations are actually calls to
'diagon'. In 'compute_dm', 'diagon' is now passed an extra (optional)
argument to signal the need to keep the eigenvectors.  The array 'psi'
will not be deallocated in diagon, but after the first true OMM step.

(Thanks to Nick Papior for finding the bug)