Converted CG comments to optimization comments, and added description of "errors" in TS
There are no code changes.

Comments in the code are changed from CG to "optimization" since Broyden or FIRE are also
enabled.

TranSiesta discussions on the extra errors printed out during the SCF for bias calculations
are now discussed thoroughly.

Also reverted the documentation details when people compile it themselves.