Removed unnecessary memory in tbtrans
There are certain lists that may easily be removed and thus reducing
memory requirements. This is in particular lists of indices that refer
to electrode regions and downfolding regions. Now, I have removed some of them
which is only minimally affecting the memory requirement, none theless they
have some effect.