Avoid the use of a non-explicitly deallocated array
In SiestaXC/mesh3d.F90 an allocatable array was not explicitly
deallocated. This is acceptable for F95 and beyond, but it could
lead to problems with some compilers or compiler options.
The array can in fact be removed by a minor change of the code.

(Thanks to Yann Pouillon and Nick Papior)