Fix 'n' quantum number for high-l polarization orbitals
When a polarization orbital had angular momentum l greater than the
rest of the the basis, its associated principal quantum number n was
off (it was set to the polarized shell's n).

This error affected the metadata in the .ion files, the wavefunction
files, and the PDOS file, so some analysis operations might have
failed to take into account those orbitals if they were requested
explicitly (e.g., a COOP calculation with 'O_3d' symbols might show
the bug; but not one with 'O_d' or 'O').