Do not get the EDM during the SCF cycle + NTPoly support
* In compute_dm, the call to the ELSI ('get_dm') interface uses a
flag to disable the generation of the EDM matrix. In this case
only the DM is computed.

After the scf loop, in 'post_scf_work', a further call to the ELSI
interface requests (only) the EDM.

Note that this is achieved through a logical flag 'Get_EDM_Only',
which is not really appropriate. It would be better to have a
more explicit flag, such as:

           output_mode="DM|EDM"

The ELSI shutdown has been moved towards the end of 'siesta_forces',
and the 'Siesta_Worker' logic preserved in that routine (even though
it is not working elsewhere).

* NTpoly support has been added to the interface
 
(Thanks to Victor M. Garcia-Suarez and Victor Yu)