Enable short-hand for entering rc information in PAO.Basis block
If the number of rc's for a given shell is less than the number of
'zetas', the program will assign the last rc value to the remaining
zetas, rather than stopping with an error.  This is particularly
useful for Bessel suites of orbitals.

A new test (Tests/bessel-rich) exemplifies this, and the previous
patches' Bessel extensions.