Clarification of eta values used in transiesta/tbtrans/phtrans calculations
In transiesta the eta value is *only* used on the real axis for the non-equilibrium
part. If the user provides an eta energy which is smaller than 0 then
the complex value of the contour is used in the self-energy calculation. This allows
one to *fine* tune the complex contour with varying complex parts and also on
a per-electrode setting.

In tbtrans the eta value is *always* used. However, it was not possible to use
the contour values. This was a limitation since a user may supply an
"equilibrium" like contour to calculate equilibrum values.
Now this is possible if specifying the eta value as a negative number.

Added more documentation on the eta values and fixed a few links to tbtrans (from TS)

Removed a deprecated default (TS.Elecs.Eta was initially set to TS.Contours.nEq.Eta).
This was bad since if one defined:
  TS.Contours.nEq 0. eV
this would mean all electrodes eta values also become 0. Which is ill-defined.

Fixed some last missing Ry**2 for phtrans in CDF files.