Clarified and fixed the complex parts of non-equilibrium contours
When using different eta values in electrodes the TSCCNEQ-* files contained
wrong imaginary parts. This commit changes the TSCCNEQ files to the following
naming scheme:

  <syslabel>.TSCCNEQ-<ID>-<mu>_<electrode>

meaning that the ID is the *internal* ID number and is useful for determining
the order of the energy-points.
The <mu> denotes that the contour is the non-equilibrium corrections to the
chemical potential named <mu>.
The <electrode> denotes that the weights in this file are associated
with electrode <electrode> and the imaginary part corresponds to the
imaginary part used in the self-energy calculation.
Previously all of these files contained the same imaginary part equal to the
TS.Contours.nEq.Eta value (and thus had no relation to the actually used eta
value for the electrode).

When the eta values are the same in all electrodes, the files will
contain the same values, however, when eta differs, so does the *imaginary*
part in the corresponding *_<electrode> files.

Conversion factors has been revised in the contour parts since it was
difficult to get the equivalent precision using external tools.
This is because the fdf-flags read in (which determines the energies)
has one conversion factor (eV->Ry) while the written files was using
the conversion factor in units.f90. This meant that there was a relative
difference of 1e-6 between the two. Now the contours are written using
the fdf-conversion factor, explicitly.

The precision in the written contour files has been increased to
maximum numerical precision. width == 17.

Added a check that all non-equilibrium electrode Eta values are not 0.

NOTE: I have asked Alberto what to do about unit conventions in the code.