Important: new output convention for the non-spin case
(work in progress)

The non-spin and full-spinor cases are similar in that they
use a single series of wavefunctions (without 'spin-channels').
However, there are approximately double the number of bands in
a given energy interval in the full-spinor case. To provide a
coherent output, the convention for the non-spin case has had
to be changed.

See the file in Util/COOP/Docs for a full account of the changes.