Initial matel enhancement by Rogeli Grima
(The following is translated from a message by Rogeli detailing the
changes so far (as of April 2, 2019):

state_init.F: I have put here the call to the initialization of the
new MATEL module (MATEL_MOD). Just before the call to Overlap. I think
its place should be another. Probably inside SIESTA_INIT.

overlap.f: I have replaced the call to new_matel with the equivalent
of the new module: get_matel_s

kinefsm.F: I have replaced the call to new_matel with the equivalent
of the new module: get_matel_t. Depending on the type of matrix the
interface is slightly different (We avoid some conditional for each
call).

nlefsm.f: I have replaced the call to new_matel with the equivalent of
the new module: get_matel_s.

naefs.F: In development. This routine also uses a type T matrix. I
think I'm going to store your data as a "special" case. 'TA'?

setup_H0.F: In this routine I have introduced a bit of timing & debug
code, which will surely disappear.

The fundamental part of the changes is in the file matel_mod.F. It is
a fortran module that contains the MATEL type. This type MATEL, tries
to be a replica of the module new_MATEL, but reducing the
dimensionality associated with the various operations ('S', 'T', 'U',
...). We give that dimensionality by creating the instances of the
MATEL type that are necessary. In this way, each operation can
dimension its structures in a personalized way. The MATEL type has
associated a series of procedures:

 * INIT (IOPER): Creates and initializes an IOPER type matrix
 * GET_MATEL (i, j, ...): Extract the element associated with the orbitals i / j.

The rest of MATEL procedures are used to build parts of the matrix,
communicate those parts between the various processes or to index the
elements of said matrix.

Apart from the MATEL type, in the module there are several routines
that are used as an interface of the type:

init_matel (): Create the matrices that are necessary
get_matel_s (): Gets an element of the matrix S.
get_matel_t (): Get an element of the matrix T.