Add interfaces for 'in place' all-reduce operations
A new module 'm_mpi_inplace' provides interfaces for the 'in place'
version of mpi_all_reduce, which was implicitly used by Rogeli Grima
in his work.

The calls might have been re-worked, but the new module will be useful
in the future to eliminate the need for extra variables and for extra
copy operations.

As the 'in-place' option is not supported by the 'classic' MPI
interface contained in mpi_siesta, the new module does not use the
latter. As side effects, the 'comm' argument is now mandatory, and
no 'MPI' timing information is provided.