Cleaned m_ts_hartree
I now know this is correct since it now deals with a 3D array.
Once in a while (when reading that code) it is easier
to decipher whether it is done correctly by doing explicit
3D arrays.
I have checked against a chain with x/y/z and
plane/elec-plane/elec-box options to assert it yields
the same result in all cases.