Added NEGF energy calculations and print-out
A minor addition to calculate NEGF-part energies.
This means that we now only calculate energies from atoms
in the density matrix elements that are updated which depends
on the options of the electrodes.
This energy calculation is *still* wrong, but it at least removes
elements that are not part of transiesta (and in particular
buffer atoms which *could* really fuck up energies).

Added a comment on energies for the documentation to let people
know that energies are not to be trusted.

All energies lives in m_energies and are prefixed with NEGF_ for clarity.