Implemented tSpin in bands/writewave and fixed a few output options
Using tSpin cleans up a bit some of the *magic* variables used
here and there. Actually the use of tSpin suggests that we need
a new variable in tSpin to clarify whether it is NCol .or. SO
which could help in certain places.

The output files are now a bit weird when comparing format
for NC.
- ioeig uses no_u * 2, 1
- bands uses no_u, 4 for WFS file
- bands uses no_u * 2, 1 for .bands
- wwave uses no_u, 4 for WFSX

It is not clear to me why they are different, and/or
if they *should* be different.
I would prefer if we can use the same convention all-over.
This may require changes all-around.

I don't have any preference. Probably I would prefer:

   no_u, spin%H

which clarifies exactly which type of calculation it was.
This is probably opposite of my previous perception.

Also, I have cleaned up print_spin which did lots of
duplicate work for SO calculations.