Fix serious bug affecting SOC calculations with ONCVPSP psml pseudos
The spin-orbit code in Siesta expects to find the lj KB projectors ordered
so that the j=l-1/2 block appears first, followed by the j=l+1/2 block.
This is the default generation order when the Siesta built-in code for
lj KB generation is used.

When reading the lj KBs from psml files, however, the required
ordering was not properly enforced. Some psml files (those generated
by 'psop' in the Siesta suite) followed the required ordering, but
some others (notably those generated by ONCVPSP, which is used as
engine for the Pseudo-Dojo database) did not. In the latter case, erroneus
results were produced.

The fix involves extracting the KB information from the PSML files in
the proper order, overriding the sequence present in the file.