Decouple computation of Delta_vscf from updates of mu bracket
In ELSI-PEXSI, the bracket for mu is updated by computing dV_min
and dV_max, the extremum shifts of the V_scf total potential.
The computation of the shifts and the bracket update have been
decoupled: the first is done wholly via a call from dhscf, and the
latter is explicitly done in the solver routines by using the
(module) variables dv_min and dv_max.

An off-by-one mistake in the bracket logs has been thus fixed.

More documentation on the procedure has been added, including a
note about the differences between the cases of mixing H and
mixing the DM.