maint: removed OCCS file
It was only meaningful for debugging Diag.WFS.Cache.
However, Siesta crashes if the number of eigenstates
is too low. Probably the *only* useful number is the
largest of the occupied states for any k-point.
And this could more trivially be written to a file.
Also, this created unnecessary IO.

Signed-off-by: Nick Papior <nickpapior@gmail.com>