bug: fixed hartree potential as pointed out by P. Ordejon and co-workers at ICN2
The problem arose because the fixation of the potential *has*
to be done at the periodic boundary. This severely disturbs the hartree
solution for capacitor like systems. But is seemingly very weak depending
on how conductive the central region is. I.e. for very conductive systems
this error did not show it self.
The ramp-central option showed in bulk-chain calculations to be better.
However, it was obviously wrong when checking the capacitor setup
(when testing these options we never did capacitors :( ).
I have removed the ramp-central option.

Another bug has been found in the weighting scheme for cases
where one of the non-equilibrium contributions was exactly 0.
This has been fixed by restructuring the calculation of the
theta variables.

Also fixed buffer calculations using TS.SolutionMethod full.

Signed-off-by: Nick Papior <nickpapior@gmail.com>