enh: charge correction at the reference energy level is now faster
Prior to this commit the Fermi correction scheme first converged
the density/hamiltonian. Then a second run with only a single
energy point was done, thus calculating the charge density at the
Fermi level.

However, this is very in-efficient since the poles are *all* at
the Fermi level. Instead of doing the second step we store the charge
at the poles and extrapolate the density to a defined eta value.

Implemented this method also showed some inconsistencies in the m_interpolate
method.

Changed TS.ChargeCorrection to TS.dQ which is shorter!

Added a file_delete method in m_os.

Still substantial clean-ups are necessary.

Signed-off-by: Nick Papior <nickpapior@gmail.com>