enh: updated dQ fermi method
Changed default values to be more conservative.

Now the TS_FERMI file will never be deleted
and it will be parsed in the code to *only*
take into account the last variables that have the same
tendency. This also means that analysis may be more
stable as we can later add mechanisms for checking
the global convergence.
The file_fermi function is now much simpler since
it won't use all dq-scf iterations. This also makes
sense since changing the Fermi-level tends to *corrupt*
the regular SCF cycle. Thus limiting the number of
previous SCF dq-cycles to take into account seems more wise.

Added a scaling of EDM when changing the fermi-level.
There is a different scaling in the electrode vs.
device region.

Added Excess charge printout in the initial
transiesta table. This is good to check initial
charge differences.

Signed-off-by: Nick Papior <nickpapior@gmail.com>