enh: performance increase for tall+skinny matrices in TS
When doing tall+skinny matrices in TS inversion it can be
speeded up by the following argument.
We want to solve this:
   A x = B

where A[n,n] and B[n,m].

When n * 2 > m the fastests algorithm is:
   GESV
otherwise when n * 2 < m, the fastests algorithm is:
   GETRF
   GETRI
   GEMM

We have accommodated these changes in the m_trimat_invert
code.
Previously we always did GESV.

Additionally when simply inverting a matrix we always
find GETRF + GETRI to be the faster one.
Regardless of whether copying data is necessary.

I have checked this with OpenBLAS and BLIS, both yield
the same result, so I guess the same will apply to MKL.

Also fixed a trim while printing the citations.

Signed-off-by: Nick Papior <nickpapior@gmail.com>