TS-maint: removed spin-dependent print-out of charges
During the SCF it ain't too important how the distribution
of charges is. So we refrain from making a spin-dependent
distribution and now instead print the full charge
distribution at the end of the SCF cycle (just after Mulliken).
Now we then print Qup - Qdn to clarify the spin configuration.

Signed-off-by: Nick Papior <nickpapior@gmail.com>