enh: fixed #40 and #42
Now Voronoi and Hirshfeld charges are
calculated per spin and printed per spin.
This causes spin-polarized calculations to
be rather difficult to parse since the
total is not printed. However, it is rather
trivially calculated.

The spin-resolved is simply based on the
Qatm / nspin splitting of spins. This may not
be the standard spin-split density but users
may correct this easily by adding/subtracting 0.5
electrons per spin channel.

Things that are changed:
- Hirshfeld method is optimized slightly.
- XML print-out to Voronoi and Hirshfeld property
- Per-spin print-out
- Increased precision of std-out (to 11.7)
  which allows for -XY.**** (maximally).
  I do not think any atom would have more than 100
  electrons (?)
- changed code to F90 and renamed file

I have also added NC and SOC implementation of both.
Instead of printing the deficiency in spin-vector it prints
the complete vector since there is no such thing as *deficiency*
spin.

However, for NC+SO I do this:
  q([1, 2]) = zvalfis(is) / 2 - q([2, 1])
Since then S will be returned with *correct* signs and Q will be
with respect zvalfis.
This may be a little weird for users since then spin is not with
respect to the atomic configuration. On the other hand the
atomic configuration is always defined as S = 0.

To clarify that charges are with respect to atomic charges the table
print-out now uses dQ.

Currently "voronoi" calculates for atom ia's Voronoi cell and for
single spin (and rhoscf is including rho_atm):

   q(ia) = - [int_\r rhoscf(\r) - rho_atm(\r)]
         = [int_\r rho_atm(\r) - rhoscf(\r)]

I have changed to the second line.

For "hirshfeld" it looks like this:

   q(ia) = zvalfis(ia) - [int_\r rhoscf(\r) * rho_atm_ia(\r) / rho_atm(\r)]

I guess this could be translated to this?:

   q(ia) = [int_\r rho_atm_ia(\r) - rhoscf(\r) * rho_atm_ia(\r) / rho_atm(\r)]

where rho_atm_ia(\r) is the atomic charge from atom ia at \r.