bug: fixed subset eigenvalue calculation for NC/SOC
This bug rarely shows it self since the occupations
are stored in a *mirror* array not used for anything but
occupations in diag*k routines.
Additionally the bug only shows it self if the number
of occupied states is greater than the *basis number
of orbitals*. I.e. for very low basis sets.