bug: fixed spin-spiral calculations
I find it hard to believe they ever worked (besides in serial). :(
Now I have fixed indices and array dimensions.

This part of the code was also exposed to the bug in #44.

I have added a test, but I have no idea what it should give.
We need some feedback on this part of the code.