bug: fixed correct weights for k-points
The fermid routine had to be changed to account for
correct spin density.
The problem arose for NC/SOC calculations and nspin = 1
in the argument list.
This resulted in weights being twice as much as they should.

This is obviously wrong.
Now the fermid routine has changed so that it is explicit
in the calculation of the correct weights for different nspinor/nspin.

I.e. one should use an argument like this:

fermid(1, 1) for non-polarized
fermid(2, 2) for polarized
fermid(2, 1) for NC/SOC

In the fermid routine we now check
nspinor == nspin == 1 and double the weights.

For all other configurations there is no doubling of
weights since the spinors take this into account.