enh: manual dipole center allowed
Moved dipole.F to F90 for sustainability.

Added a get_dipole_center method which
checks for Dipole.Center block which should contain

 x y z <coord-unit Ang or Bohr>

which will then be used as the default location of the
dipole. If the block does not exist, the regular
average position is used.

The reason for this is that some molecules do not have
the dipole center at the center of the molecule.
In such cases it is vital that the users specifies
the dipole centers, and if a MD/CG run is performed
retain the center-of the dipole where it is, e.g.
using advanced constraints.

This is the only deficiency of this, the center is fixed
for all MD simulations, if Dipole.Center is specified.
We should consider other options, e.g. Center-Of-Positions

%block Dipole.Center
  atom 1 2 4
  atom 5 5 5
%endblock

Which would take the center of positions like this:

  sum(xa(:, [1,2,4,5,5,5]), 2) / 6

the reason for allowing multiple indices is to counter another
site of a molecule, e.g.

            O2
           /
  O - X - X
          \\
	    O

One could double the weight of the lower O (fictive example).