enh: added vacuum dipole correction
Now one can calculate the vacuum dipole from 2 methods:

1. The original dipole method (now also with a center).
2. By looking at a vacuum region and determining the electric
   field from the FFT solution of the Poisson equation.

In all *regular* calculations these are equivalent.

However, when one uses gates one have to be really careful
on how to calculate the true dipole that counters the system.
This is because a charged gate does not enter in the calculation
of the center of coordinates and thus one can easily get *wrong*
dipoles. The charged gate ensures there is a field between the gate
and the system. But the dipole position will only be taking into account
the system and thus the contribution of the gate will have a much larger
weight.

This commit fixes this by allowing one to choose a vacuum point where
one wishes the electric field to be exactly 0.
The method is simply doing numerical differentiation along the given dipole
direction (currently only 1 direction is allowed). In the pure vacuum
this should be a straight line.
To decipher when one is not in the vacuum region we use a threshhold in
the change of the electric field. I.e. if the electric field changes by more than
the tolerance the line is considered to be parabolic and we are not in a vacuum region.
Another reason for using this is that it is agnostic about the placement of the dipole.
It will correct the *global* induced dipole that the FFT adds to the potential.

The actual field calculated is the minimum electric field in the
region of the vacuum.
We choose the minimum since that is stable and should be more stable than
calculating the average (which then depends on the tolerance).

I have added a test which shows how this may be used. h2o_2_dipol_gate
Where if one changes to "Dipole charge" one will get the wrong results.