enh: enabled mesh distributions to have different nsm
This should hopefully allow one to convert a sequential
distribution to another sequential distribution with
a different mesh subdivisions ordering.

This won't work for clustered distributions (packed
mesh sub divisions) since then the chances of there
being a non-full mesh sub division is pretty big.
This *could* be handled by duplicating the boundary elements
but requires quite a bit of logic.

Since now nsm is part of the distribution class we don't
need it as an argument to the setMeshDistr and friends.