maint: changed default for electrode bulk
while the electrode bulk DM option really helps for
correctly setup systems (proper screening) it also
is very problematic for not fully optimal systems.
This is because the DM in the electrode region may be
a little too far from the proper bulk DM.

This is even worse for big systems since the SCF is
very sensitive. As such I have decided to revert this
option to its less attractive default. :(

Clarified documentation here and there.