enh: improved OpenMP implementation of Bloch-expansions
After IWOMP 2020 attendence I was encouraged to play
with different versions of Bloch-expansions using different
OpenMP mechanisms.

I mainly played with the single-matrix expansion algorithm
which I think is a good testing grounds.
From this experience I extended the 2 other methods (HS and HS+G)
to use the one that seems to be best.

The following algorithms are implemented:

- many omp do statements (the former implementation)
  This will for every omp do do a scheduling of the
  loop and would infer some overhead as the iteration
  count is determined for each construct.

- a task + dependency construct. This is by far the easiest
  implementation since each sub-matrix calculated once the prior
  dependencies are calculated. In this implementation one does
  not care about the order of the loop constructs as task
  dependencies takes care of correct order.
  As such this is much simpler to parallelize.
  One note is that dependencies should be as minimal as possible

  E.g. one should NOT do depend(inout:M(:,1,:,1)), rather
  depend(inout:M(1,1,1,1)) suffices since all elements
  are calculated with the same dependency scheme.
  This is only valid for full array-section dependencies, always
  keep a minimum dependency list.

- A manual loop division. Here we once calculate one of the matrix
  dimensions thread-splitting and do all without any dependencies.

  This has 2 advantages:

  1. contrary to omp do, this only calculates the splitting once
  2. since one thread is only calculating on some columns, there
     is no dependency and thus no barriers are necessary.

- A workshare division. This is extremely simple to implement,
  but relies on workshare directives which typically gives
  inferior parallelization.

My findings show that the manual implementation is the fastest
in most cases. The task construct is very fast as well and
seems best for very large matrices and bloch-factors, but very
little indeed.

The current choice is:
If compiled with openmp -> manual
If not compiled with openmp -> do

I also renamed the function calls to better match their intent.

Now the bloch_unfold.F90 file is a complete program when compiled
like this:

 -D__BLOCH_UNFOLD_M_TEST

and optionally with OpenMP. Then one can run it with several
options for checking the implementations and doing benchmarks.